split issue show page spec into separate file
spec/features/issues_spec is too long and unwieldy
